requirements.txt
flask==3.0.3
flask-cors==4.0.0
PyMySQL==1.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-dotenv==1.0.0
# openai depends on jiter (JSON parsing); missing/broken jiter causes: No module named 'jiter.jiter'
jiter>=0.8.0
openai>=1.40.0
